port models

+++ src/turbomind/models/llama/llama_kernels.cu
@@ -112,7 +112,7 @@ template<typename T>
 void invokeAddResidual(T* out, const T* in, int m, int n, cudaStream_t stream)
 {
     auto total = static_cast<size_t>(m) * n;
-    dim3 block(std::min(total, 1024UL));
+    dim3 block(std::min((unsigned long)total, 1024UL));
     dim3 grid((total + block.x - 1) / block.x);
 
     addResidual<<<grid, block, 0, stream>>>(out, in, total);
